PMM-7259 separate IntegratedAlertingTabs
