+++ public/app/features/integrated-alerting/IntegratedAlertingPage.tsx
@@ -1,58 +1,12 @@
-import React, { FC, useMemo, useState, useEffect } from 'react';
-import { TabsBar, TabContent, Tab, useStyles, Spinner } from '@grafana/ui';
+import React, { FC, useState, useEffect } from 'react';
+import { useStyles, Spinner } from '@grafana/ui';
 import { Messages } from './IntegratedAlerting.messages';
 import { getStyles } from './IntegratedAlerting.styles';
-import { TabKeys } from './IntegratedAlerting.types';
-import { Alerts, AlertRuleTemplate, AlertRules, NotificationChannel } from './components';
 import { IntegratedAlertingService } from './IntegratedAlerting.service';
 import { logger } from '@percona/platform-core';
 import { EmptyBlock } from './components/EmptyBlock';
 import { PMM_SETTINGS_URL } from './IntegratedAlerting.constants';
-
-const IntegratedAlertingTabs: FC = () => {
-  const [activeTab, setActiveTab] = useState(TabKeys.alerts);
-  const tabs = useMemo(
-    () => [
-      {
-        label: Messages.tabs.alerts,
-        key: TabKeys.alerts,
-        component: <Alerts key={TabKeys.alerts} />,
-      },
-      {
-        label: Messages.tabs.alertRules,
-        key: TabKeys.alertRules,
-        component: <AlertRules key={TabKeys.alertRules} />,
-      },
-      {
-        label: Messages.tabs.alertRuleTemplates,
-        key: TabKeys.alertRuleTemplates,
-        component: <AlertRuleTemplate key={TabKeys.alertRuleTemplates} />,
-      },
-      {
-        label: Messages.tabs.notificationChannels,
-        key: TabKeys.notificationChannels,
-        component: <NotificationChannel key={TabKeys.notificationChannels} />,
-      },
-    ],
-    []
-  );
-
-  return (
-    <>
-      <TabsBar>
-        {tabs.map(tab => (
-          <Tab
-            key={tab.key}
-            label={tab.label}
-            active={tab.key === activeTab}
-            onChangeTab={() => setActiveTab(tab.key)}
-          />
-        ))}
-      </TabsBar>
-      <TabContent>{tabs.find(tab => tab.key === activeTab).component}</TabContent>
-    </>
-  );
-};
+import { IntegratedAlertingTabs } from './components/IntegratedAlertingTabs';
 
 export const IntegratedAlertingContent: FC<{ loadingSettings: boolean; alertingEnabled: boolean }> = ({
   loadingSettings,
