+++ public/app/features/integrated-alerting/components/IntegratedAlertingTabs/index.ts
@@ -1,0 +1,1 @@
+export * from './IntegratedAlertingTabs';
